requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.21.0
Flask==3.0.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1